fix:merge 이후 발생한 버그 수정 GameDrawManager 수정
